requirements.txt
asgiref==3.8.1
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.12.2
